+++ lib/ui/transit/google_sheet/export_order_view.dart
@@ -1,35 +1,27 @@
 import 'package:flutter/material.dart';
-import 'package:possystem/components/meta_block.dart';
 import 'package:possystem/components/sign_in_button.dart';
 import 'package:possystem/components/style/snackbar.dart';
 import 'package:possystem/components/style/snackbar_actions.dart';
 import 'package:possystem/constants/constant.dart';
-import 'package:possystem/constants/icons.dart';
 import 'package:possystem/helpers/logger.dart';
-import 'package:possystem/helpers/util.dart';
 import 'package:possystem/models/repository/seller.dart';
 import 'package:possystem/translator.dart';
 import 'package:possystem/ui/transit/exporter/google_sheet_exporter.dart';
-import 'package:possystem/ui/transit/transit_order_list.dart';
-import 'package:possystem/ui/transit/transit_order_range.dart';
+import 'package:possystem/ui/transit/google_sheet/spreadsheet_dialog.dart';
+import 'package:possystem/ui/transit/order_widgets.dart';
 import 'package:possystem/ui/transit/widgets.dart';
-
-import 'order_setting_page.dart';
-import 'spreadsheet_selector.dart';
-
-const _cacheKey = 'exporter_order_google_sheet';
 
 class ExportOrderView extends StatefulWidget {
   final ValueNotifier<DateTimeRange> ranger;
-
-  final ValueNotifier<String> stateNotifier;
-
+  final ValueNotifier<OrderSpreadsheetProperties> properties;
+  final TransitStateNotifier stateNotifier;
   final GoogleSheetExporter exporter;
 
   const ExportOrderView({
     super.key,
     required this.ranger,
     required this.stateNotifier,
+    required this.properties,
     required this.exporter,
   });
 
@@ -38,9 +30,6 @@ class ExportOrderView extends StatefulWidget {
 }
 
 class _ExportOrderViewState extends State<ExportOrderView> {
-  final selector = GlobalKey<SpreadsheetSelectorState>();
-  late OrderSpreadsheetProperties properties;
-
   @override
   Widget build(BuildContext context) {
     return TransitOrderList(
@@ -48,132 +37,112 @@ class _ExportOrderViewState extends State<ExportOrderView> {
       formatOrder: (order) => OrderTable(order: order),
       memoryPredictor: memoryPredictor,
       warning: S.transitGSOrderMetaMemoryWarning,
-      leading: Column(
-        children: [
-          Padding(
-            padding: const EdgeInsets.fromLTRB(kHorizontalSpacing, kTopSpacing, kHorizontalSpacing, kInternalSpacing),
-            child: SignInButton(
-              signedInWidget: SpreadsheetSelector(
-                key: selector,
-                notifier: widget.stateNotifier,
-                exporter: widget.exporter,
-                cacheKey: _cacheKey,
-                existLabel: S.transitGSSpreadsheetExportExistLabel,
-                existHint: S.transitGSSpreadsheetExportExistHint,
-                emptyLabel: S.transitGSSpreadsheetExportEmptyLabel,
-                emptyHint: S.transitGSSpreadsheetExportEmptyHint(S.transitGSSpreadsheetOrderDefaultName),
-                fallbackCacheKey: 'exporter_google_sheet',
-                defaultName: S.transitGSSpreadsheetOrderDefaultName,
-                requiredSheetTitles: requiredSheetTitles,
-                onPrepared: exportData,
-              ),
-            ),
+      leading: Padding(
+        padding: const EdgeInsets.fromLTRB(14.0, kTopSpacing, 14.0, kInternalSpacing),
+        child: SignInButton(
+          signedInWidget: TransitOrderExportHead(
+            title: '網路匯出',
+            subtitle: '注意，由於 Google 的限流，有時會無法成功送出，需多次嘗試。\n建議大資料可以透過 Excel 或 CSV 匯出。',
+            trailing: const Icon(Icons.upload_file_outlined),
+            ranger: widget.ranger,
+            properties: widget.properties,
+            onTap: _export,
           ),
-          TransitOrderRange(notifier: widget.ranger),
-          ListTile(
-            key: const Key('edit_sheets'),
-            title: Text(S.transitGSOrderSettingTitle),
-            subtitle: MetaBlock.withString(
-              context,
-              [
-                S.transitGSOrderMetaOverwrite(properties.isOverwrite.toString()),
-                S.transitGSOrderMetaTitlePrefix(properties.withPrefix.toString()),
-                // This message may break the two lines limit, so put it at the end.
-                properties.requiredSheets.map((e) => e.name).join('、'),
-              ],
-              maxLines: 2,
-            ),
-            isThreeLine: true,
-            trailing: const SizedBox(
-              height: double.infinity,
-              child: Icon(KIcons.edit),
-            ),
-            onTap: editSheets,
-          ),
-        ],
+        ),
       ),
     );
   }
 
-  @override
-  void initState() {
-    super.initState();
-    properties = OrderSpreadsheetProperties.fromCache();
+  Future<void> _export(BuildContext context) async {
+    widget.stateNotifier.exec(
+      () => showSnackbarWhenFutureError(
+        _startExport(),
+        'excel_export_failed',
+        context: context,
+      ).then((link) {
+        if (link != null && link.isNotEmpty) {
+          Log.out('export finish', 'gs_export');
+          showSnackBar(
+            S.actSuccess,
+            // ignore: use_build_context_synchronously
+            context: context,
+            action: LauncherSnackbarAction(
+              label: S.transitGSSpreadsheetSnackbarAction,
+              link: link,
+              logCode: 'gs_export',
+            ),
+          );
+        }
+      }),
+    );
   }
 
-  Map<SheetType, String> requiredSheetTitles() {
-    final prefix = properties.withPrefix ? '${widget.ranger.value.formatCompact(S.localeName)} ' : '';
+  /// Export all data to spreadsheet.
+  ///
+  /// 1. Ask user to select a spreadsheet.
+  /// 2. Prepare the spreadsheet, make all sheets ready.
+  /// 3. Export data to the spreadsheet.
+  Future<String?> _startExport() async {
+    // Step 1
+    GoogleSpreadsheet? ss = await SpreadsheetDialog.show(
+      context,
+      exporter: widget.exporter,
+      cacheKey: importCacheKey,
+      fallbackCacheKey: exportCacheKey,
+    );
+    if (ss == null || !mounted) {
+      return '';
+    }
 
-    return {
-      for (final sheet in properties.requiredSheets)
-        SheetType.values.firstWhere((e) => e.name == sheet.type.name): '$prefix${sheet.name}',
-    };
-  }
+    // Step 2
+    final sheetTitles = widget.properties.value.parseTitles(widget.ranger.value);
+    final ables = sheetTitles.keys.toList();
+    final titles = sheetTitles.values.toList();
+    final sheets = ss.sheets.where((e) => titles.contains(e.title)).toList();
+    ss = await prepareSpreadsheet(
+      context: context,
+      exporter: widget.exporter,
+      stateNotifier: widget.stateNotifier,
+      defaultName: S.transitGSSpreadsheetModelDefaultName,
+      cacheKey: exportCacheKey,
+      sheets: titles,
+      spreadsheet: ss,
+    );
+    if (ss == null || !mounted) {
+      return '';
+    }
 
-  /// [SpreadsheetSelector] validate the basic data before actually exporting.
-  Future<void> exportData(
-    GoogleSpreadsheet ss,
-    Map<SheetType, GoogleSheetProperties> prepared,
-  ) async {
-    widget.stateNotifier.value = S.transitGSProgressStatusFetchLocalOrders;
+    // Step 3
+    Log.ger('gs_import', {'spreadsheet': ss.id, 'sheets': titles});
+
     final orders = await Seller.instance.getDetailedOrders(
       widget.ranger.value.start,
       widget.ranger.value.end,
     );
-    Log.ger('gs_export', {'spreadsheet': ss.id, 'target': 'order'});
-
-    final data = prepared.keys.map((e) => e.orderFormattable).map((able) => orders.expand((order) {
+    final data = ables.map((able) => orders.expand((order) {
           return able.formatRows(order).map((l) {
             return l.map((v) => v.value).toList();
           });
         }));
 
-    if (properties.isOverwrite) {
+    if (widget.properties.value.isOverwrite) {
       widget.stateNotifier.value = S.transitGSProgressStatusOverwriteOrders;
       await widget.exporter.updateSheetValues(
         ss,
-        prepared.values,
+        sheets,
         data,
-        prepared.keys.map((key) => key.orderFormattable.formatHeader()),
+        ables.map((able) => able.formatHeader()),
       );
-    } else {
-      final it = data.iterator;
-      for (final entry in prepared.entries) {
-        it.moveNext();
-        final name = S.transitModelName(entry.key.name);
-        widget.stateNotifier.value = S.transitGSProgressStatusAppendOrders(name);
-        await widget.exporter.appendSheetValues(ss, entry.value, it.current);
-      }
+
+      return ss.toLink();
     }
 
-    Log.out('export finish', 'gs_export');
-    if (mounted) {
-      showSnackBar(
-        S.actSuccess,
-        context: context,
-        action: LauncherSnackbarAction(
-          label: S.transitGSSpreadsheetSnackbarAction,
-          link: ss.toLink(),
-          logCode: 'gs_export',
-        ),
-      );
+    for (final (i, rows) in data.indexed) {
+      widget.stateNotifier.value = S.transitGSProgressStatusAppendOrders(ables[i].l10nValue);
+      await widget.exporter.appendSheetValues(ss, sheets[i], rows);
     }
-  }
 
-  void editSheets() async {
-    final other = await showAdaptiveDialog<OrderSpreadsheetProperties>(
-      context: context,
-      builder: (context) => OrderSettingPage(
-        properties: properties,
-        sheets: selector.currentState?.spreadsheet?.sheets,
-      ),
-    );
-
-    if (other != null && mounted) {
-      setState(() {
-        properties = other;
-      });
-    }
+    return ss.toLink();
   }
 
   /// These values are based on the actual data:
